Added a common `SideNavigationLayout` component Used this common component in `/dashboard` & `/profile` Added `onAccountsChanged()` in `@basketo/web-utils`
